requirements.txt
Flask
PyPDF2
gunicorn
pdf2image
Pillow
supabase
google-cloud-storage
